fix: refatorando service para adequar ao  async/await

+++ src/cli/CategoryService.js
@@ -1,12 +1,9 @@
-
-
-
 const URL_SERVER = 'http://127.0.0.1:3000/categories';
 
 function objetoEstaVazio(objeto) {
     return (Object.entries(objeto).length === 0) ? true : false;
-
 }
+
 async function gerarNovoId() {
     const listaDeCategorias = (await CategoryService.findCategories()).resultado;
     let maiorId = listaDeCategorias[0].id;
@@ -21,8 +18,8 @@ export default class CategoryService {
     // METODOS GET
     static async findCategories() {
         try {
-            const listaDeCategorias = await fetch(URL_SERVER, { method: "GET" })
-                .then(response => response.json());
+            const response = await fetch(URL_SERVER, { method: "GET" });
+            const listaDeCategorias = await response.json();
             return { statusCode: 200, resultado: listaDeCategorias };
 
         } catch (error) {
@@ -31,32 +28,34 @@ export default class CategoryService {
     }
 
     static async findCategoryById(id) {
-        const url = `${URL_SERVER}/${id}`
+        const url = `${URL_SERVER}/${id}`;
         try {
-            const categoriaEncontrada = await fetch(url, { method: "GET" })
-                .then(response => response.json());
+            const response = await fetch(url, { method: "GET" });
+            const categoriaEncontrada = await response.json();
 
             return (!objetoEstaVazio(categoriaEncontrada) ? {
                 statusCode: 200, resultado: categoriaEncontrada
             } : {
                 statusCode: 404, resultado: "categoria não encontrada"
             });
+
         } catch (error) {
             return { statusCode: 500, resultado: (`Ocorreu um erro no servidor ${error}`) };
         }
-
     }
 
     // METODO POST
     static async createCategory(novaCategoria) {
         try {
             novaCategoria.id = await gerarNovoId();
-            const categoriaInserida = await fetch(URL_SERVER, {
+            const response = await fetch(URL_SERVER, {
                 method: "POST",
                 body: JSON.stringify(novaCategoria),
                 headers: { "Content-type": "application/json; charset=UTF-8" }
-            }).then(response => response.json());
+            });
+            const categoriaInserida = await response.json();
             return { statusCode: 201, resultado: categoriaInserida };
+
         } catch (error) {
             return { statusCode: 500, resultado: (`Ocorreu um erro no servidor ${error}`) };
         }
@@ -64,33 +63,43 @@ export default class CategoryService {
 
     // METODO PUT
     static async updateCategory(id, categoria) {
-        const url = `${URL_SERVER}/${id}`
+        const categoriaStatus = (await this.findCategoryById(id)).statusCode;
+        if (categoriaStatus === 404) {
+            return { statusCode: 404, resultado: "categoria não encontrada" }
+        }
+
+        const url = `${URL_SERVER}/${id}`;
         try {
-            const categoriaAtualidada = await fetch(url, {
+            const response = await fetch(url, {
                 method: 'PUT',
                 body: JSON.stringify(categoria),
                 headers: { "Content-type": "application/json; charset=UTF-8" }
-            }).then(response => response.json())
+            });
 
+            const categoriaAtualidada = await response.json();
             return { statusCode: 200, resultado: categoriaAtualidada };
 
         } catch (error) {
             return { statusCode: 500, resultado: (`Ocorreu um erro no servidor ${error}`) };
-
         }
     }
     // METODO DELETE
     static async deleteCategory(id) {
-        const url = `${URL_SERVER}/${id}`
+        const categoriaStatus = (await this.findCategoryById(id)).statusCode;
+        if (categoriaStatus === 404) {
+            return { statusCode: 404, resultado: "categoria não encontrada" }
+        }
+
+        const url = `${URL_SERVER}/${id}`;
         try {
-            const categoriaExcluida = await fetch(url, {
+            const response = await fetch(url, {
                 method: 'DELETE'
-            }).then(response => response.json())
+            });
+            const categoriaExcluida = await response.json();
+
             return { statusCode: 200, resultado: categoriaExcluida };
         } catch (error) {
             return { statusCode: 500, resultado: (`Ocorreu um erro no servidor ${error}`) };
         }
     }
-
-
 }
